experience permissions

+++ app/Http/Controllers/ExperienceController.php
@@ -11,6 +11,7 @@ class ExperienceController extends Controller
 {
     public function create(Character $character)
     {
+        $this->authorize('updateCharsheetGm', $character);
         return view('experiences.create', [
             'character' => $character
         ]);
@@ -18,6 +19,7 @@ class ExperienceController extends Controller
 
     public function store(ExperienceRequest $request, Character $character)
     {
+        $this->authorize('updateCharsheetGm', $character);
         $validated = $request->validated();
         $character->experiences()->create($validated);
 
@@ -26,6 +28,7 @@ class ExperienceController extends Controller
 
     public function edit(Character $character, Experience $experience)
     {
+        $this->authorize('update', $character);
         return view('experiences.edit', [
             'character' => $character,
             'experience' => $experience
@@ -34,6 +37,7 @@ class ExperienceController extends Controller
 
     public function update(ExperienceRequest $request, Character $character, Experience $experience)
     {
+        $this->authorize('update', $character);
         $validated = $request->validated();
         $experience->update($validated);
 
@@ -42,6 +46,7 @@ class ExperienceController extends Controller
 
     public function destroy(Character $character, Experience $experience)
     {
+        $this->authorize('update', $character);
         $experience->delete();
 
         return redirect()->route('characters.show', $character);
@@ -49,6 +54,7 @@ class ExperienceController extends Controller
 
     public function setView(Character $character, Experience $experience)
     {
+        $this->authorize('updateCharsheetGm', $character);
         return view('experiences.set', [
             'character' => $character,
             'experience' => $experience
@@ -57,6 +63,7 @@ class ExperienceController extends Controller
 
     public function set(Request $request, Character $character, Experience $experience)
     {
+        $this->authorize('updateCharsheetGm', $character);
         $validated = $request->validate([
             'level' => ['min:1', 'max:10']
         ]);
